+++ LMIPy/__init__.py
@@ -1,4 +1,8 @@
-from .lmipy import Collection, Dataset, Layer, Table, Vocabulary, Metadata
+from .lmipy import Vocabulary, Metadata
+from .layer import Layer
+from .dataset import Dataset
+from .collection import Collection
+from .table import Table
 from pkg_resources import get_distribution
 
 __version__ = get_distribution('LMIPy').version
