+++ setup.py
@@ -5,7 +5,7 @@ with open("README.md", "r") as fh:
 
 setuptools.setup(
     name="LMIPy",
-    version="0.0.3",
+    version="0.0.4",
     author="Vizzuality",
     author_email="benlaken@icloud.com",
     description="Interface to data and layers in the Resource Watch API",
